#9579
Add SQLR_PRINTF_FORMAT to the raise*Event() declarations in
sqlrserver.h and their counterparts in sqlrlistener.h, so a non-literal
or mismatched-arity format argument is caught at compile time - the
same gap that let #9575 and #9578 slip through by hand. Wired
-Wformat-nonliteral/-Wformat-security into the build via a new
configure.in probe (mirroring the existing -Wall check, degrading to
nothing on an unsupporting compiler; no rudiments changes needed).

Fixed the resulting fallout: 11 pre-existing call sites that pass a
locally-rendered, non-client-controlled string as a format argument
(postgresql.cpp, sqlrclient.cpp, firebird.cpp, oracle.cpp), plus one
previously-unknown bug the new annotations surfaced on their own -
debugHexDump() passed a size_t precision argument where "%.*s" expects
int, undefined behavior in the varargs read.

+++ src/server/sqlrservermodule.cpp
@@ -98,7 +98,7 @@ void sqlrservermodule::debugHexDump(const byte_t *data, uint64_t size) {
 	pvt->_logbuffer.clear();
 	pvt->_logbuffer.printHex(data,size,0);
 	cont->raiseDebugWriteEvent("%.*s",
-			pvt->_logbuffer.getSize(),
+			(int)pvt->_logbuffer.getSize(),
 			pvt->_logbuffer.getString());
 }
